Move the products API utilities to a shared folder
